删除没用的

+++ MicroCommunity/Support/ITTCategory/ObjAddtions/NSDate+ITTAdditions.m
@@ -197,7 +197,7 @@
     NSCalendar *calendar = [NSCalendar currentCalendar];
     NSUInteger unitFlags = NSYearCalendarUnit | NSMonthCalendarUnit | NSDayCalendarUnit | NSHourCalendarUnit | NSMinuteCalendarUnit | NSSecondCalendarUnit;
     NSDateComponents *dateComponent = [calendar components:unitFlags fromDate:now];
-    int year = [dateComponent year];
+    NSInteger year = [dateComponent year];
     return year;
 }
 @end
